fix: library multiselect not accounting for published state
